(#136) BaseViewModel 로 현재위치 가져오는 것 이동
