+++ app/src/main/java/com/boostcamp/travery/useraction/list/UserActionListViewModel.kt
@@ -1,16 +1,10 @@
 package com.boostcamp.travery.useraction.list
 
-import android.Manifest
 import android.app.Application
-import android.content.Context
-import android.content.pm.PackageManager
 import android.location.Location
-import android.location.LocationManager
-import androidx.core.content.ContextCompat
 import androidx.databinding.ObservableArrayList
 import androidx.lifecycle.MutableLiveData
 import com.boostcamp.travery.Injection
-import com.boostcamp.travery.MyApplication
 import com.boostcamp.travery.base.BaseViewModel
 import com.boostcamp.travery.data.model.UserAction
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -48,27 +42,4 @@ class UserActionListViewModel(application: Application) : BaseViewModel(applicat
             curLocation.value = it
         }
     }
-
-    private fun getLastKnownLocation(): Location? {
-        val fineLocPerm = ContextCompat.checkSelfPermission(getContext(), Manifest.permission.ACCESS_FINE_LOCATION)
-        val courseLocPerm = ContextCompat.checkSelfPermission(getContext(), Manifest.permission.ACCESS_COARSE_LOCATION)
-
-        val locationManager = getContext().getSystemService(Context.LOCATION_SERVICE) as LocationManager
-        val providers = locationManager.getProviders(true)
-        var bestLocation: Location? = null
-
-        val isGranted = PackageManager.PERMISSION_GRANTED
-        if (fineLocPerm == isGranted && courseLocPerm == isGranted) {
-            for (provider in providers) {
-                val mLocation = locationManager.getLastKnownLocation(provider) ?: continue
-                if (bestLocation == null || mLocation.accuracy < bestLocation.accuracy) {
-                    bestLocation = mLocation
-                }
-            }
-        }
-
-        return bestLocation
-    }
-
-    private fun getContext() = getApplication<MyApplication>()
 }
